Stock Update completed

+++ aynastock/src/main/java/com/ayna/aynastock/repo/StockRepo.java
@@ -7,5 +7,6 @@ import com.ayna.aynastock.entity.Stock;
 
 @Repository
 public interface StockRepo extends JpaRepository<Stock, Long> {
-	Stock findByProductId(int productId);
+	Stock findByProductId(Long productId);
+	
 }
